拆分 DownloaderServer
