+++ src/main/java/com/ghostchu/peerbanhelper/event/PBHServerStartedEvent.java
@@ -1,11 +1,11 @@
 package com.ghostchu.peerbanhelper.event;
 
-import com.ghostchu.peerbanhelper.PeerBanHelperServer;
+import com.ghostchu.peerbanhelper.PeerBanHelper;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 
 @Data
 @AllArgsConstructor
 public final class PBHServerStartedEvent {
-    private PeerBanHelperServer instance;
+    private PeerBanHelper instance;
 }
