+++ src/main/java/com/ghostchu/peerbanhelper/gui/impl/GuiImpl.java
@@ -1,6 +1,6 @@
 package com.ghostchu.peerbanhelper.gui.impl;
 
-import com.ghostchu.peerbanhelper.PeerBanHelperServer;
+import com.ghostchu.peerbanhelper.PeerBanHelper;
 import com.ghostchu.peerbanhelper.gui.ProgressDialog;
 import com.ghostchu.peerbanhelper.gui.TaskbarControl;
 import org.jetbrains.annotations.Nullable;
@@ -16,7 +16,7 @@ public interface GuiImpl {
 
     void close();
 
-    default void onPBHFullyStarted(PeerBanHelperServer server) {
+    default void onPBHFullyStarted(PeerBanHelper server) {
     }
 
     void createNotification(Level level, String title, String description);
